lexer: two-character operator tokens and change tokens include `start` and `end` instead of `length`
I wanted a preprocessor in converting the tokens into their appropriate
two-character operator tokens using a separate token struct but similar
to how the rust compiler does it but I think that's a little too much.
Also, I wanted the separate token struct to include the `start` and
`end` data.