implement `TokenStream`
This is a high-level interface for the lexer. It converts the lexical
tokens to parse tokens where it will make more sense to the parser in
building the AST. It also calculates the start and end positions of the
tokens as well as skip whitespace.

`TokenStream` is "on-demand." This means that it doesn't store the parse
tokens in memory and only creates the parse tokens when asked.

I also still need to implement a "peek" function where we can see the
token it's currently on, but doesn't consume it.