parser + tokenstream: big commit
The statements side of the parser works for the most part. I still need
to fix the `parse_params()` function. It emits an error if there's only
one parameter and doesn't have a comma at the end.

I also changed the tokenstream to just be a `Vec<ast::Token>`. I'm tired
of trying to make it work. I think with this set up though, I can change
things around a little to make it work. However, I will do that later. I
changed the old `TokenStream` struct into a `StringReader` and it just
uses the same logic to tokenize things. It's a private struct that's
only used to create `TokenStream`s.