resolve #3
Instead of purely removing operator groupings, I modeled it after the
Rust compiler and just have a `Op` group that have an `OpKind` with it.
The operators in the `OpKind` can be also be used for `OpEq`
representing compound assignments. This implies that only arithmetic and
bitwise operators are in this `Op` group since there's not compound
assignment equivalent for boolean and comparison operators.

i think i'm copying the rust compiler a little too much but i think it's
alright